wyodrebnienie PlayerUI do osobnej klasy
